export hook with addToast in it
